Move AuthFilter to a separate file
